refactor sqlfs_proc_write to eliminate calloc/memcpy for almost all writes
The only time we need to calloc a new buf and then memcpy to it is when the
offset is past the end of the existing file (EOF). Then we need to fill the
space between EOF and offset with 0s, then write buf.  set_value() then
writes that whole thing.

Other cases, we just pass buf directly to read from.  This saves all of the
calloc/memcpy/free/clean_value activity for the vast majority of writes.

refs #553