remove confusing comma syntax in if statements
There is a obscure bit of valid C syntax where you can run a statement in an
if statement, then put a comma, then the actual test.  The result of the part
before the comma is then ignored by the if statement. This is ugly and
unreadable, so I removed it.

fixes #1726