refactor get_value() to write directly into the buffer given to it
Before, this code was allocating lots of memory, and making multiple copies
of things.  It would allocate chunks of memory the size of the entire file
no matter how small the read.  There is not need for such silliness since
a buffer and size is handed to sqlfs_proc_read() and sqlfs_proc_readlink(),
the two main users of get_value().  So this refactors things to just put
the data directly into that buffer as it is being read.