add separate APIs for working with passwords and raw keys
If people are going to handle the password themselves, or for use as a FUSE
filesystem, then most likely, passwords will be the way to go.  For use in
IOCipher in conjunction with CacheWord, things work best if raw key data is
used.  That way CacheWord can pass the raw key data that it manages
directly to IOCipher.

fixes #3173 https://dev.guardianproject.info/issues/3173