base open/close functions on the init functions
In order to use the open/close mode as a way to track the mounted state
of the filesystem, the open/close mode needs to work with the init aka
threaded mode.  That means that the sqlfs instance created by sqlfs_open*
must be assigned to the pthread_key and the filename and password must be
properly cached for any threads that might be created.

refs #3402 https://dev.guardianproject.info/issues/3402